requirements.txt
# Minimal dependencies for THIS repo (Flask UI + Admin ingestion + Qdrant)
#
# NOTE: The previous (very large) requirements list included numpy/langchain/etc.
# On Windows + Python 3.14 that forced a numpy source build and blocked installs.
# This project does not require numpy/langchain to run.

Flask==3.0.3
qdrant-client==1.17.1
python-dotenv==1.2.2

# Admin document upload/conversion (optional but recommended)
pypdf==6.0.0
python-docx==1.2.0
Pillow==11.3.0
pytesseract==0.3.13

# Optional (heavy) for higher-quality embeddings:
# sentence-transformers==5.4.0
